# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OneAce API
  version: 1.0.0
  description: |
    OneAce is a comprehensive inventory management and stock counting platform.

    ## Authentication
    Most endpoints require authentication via session cookie (from Better Auth).
    Some endpoints (webhooks, cron, health) are public but have additional security measures.

    ## Rate Limiting
    Rate limits vary by endpoint. When rate-limited, the response includes:
    - `Retry-After` header: seconds until the next request is allowed
    - `X-RateLimit-*` headers: current limit status

    ## Error Responses
    All error responses include an `error` field describing the issue.
    Common status codes:
    - 400: Invalid request
    - 401: Unauthenticated or invalid credentials
    - 403: Forbidden / insufficient permissions
    - 404: Resource not found
    - 429: Rate limited
    - 500: Internal server error
    - 503: Service unavailable
servers:
  - url: https://oneace-next-local.vercel.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local development
...
paths:
...
  /labels/custom/pdf:
    post:
      tags:
        - Labels
      summary: Generate custom labels PDF
      description: |
        Generate custom labels from template for items, bins, or warehouses.
        Returns JSON (PDF generation is placeholder in implementation).
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                templateId:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [warehouse, bin, item]
                      id:
                        type: string
                  minItems: 1
              required:
                - templateId
                - items
      responses:
        200:
          description: Label data (JSON placeholder)
          content:
            application/json:
              schema:
                type: object
        400:
          description: Invalid request
        404:
          description: Template or items not found
        401:
          description: Unauthenticated
  /onboarding/organization:
    post:
      tags:
        - Onboarding
      summary: Create organization
      description: |
        Create a new organization with authenticated user as OWNER.
        Automatically creates default warehouse and sends welcome email.
        Rate limited: 3 per hour per user.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 80
                  description: Organization name
              required:
                - name
      responses:
        201:
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      slug:
                        type: string
                      warehouses:
                        type: array
        400:
          description: Invalid request
        401:
          description: Unauthenticated
        429:
          description: Rate limit exceeded
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Better Auth session cookie
    bearerToken:
      type: http
      scheme: bearer
      description: Bearer token for cron endpoints (CRON_SECRET)